requirements.txt
flask>=3.0
google-genai>=1.0
lz4>=4.0
psycopg2-binary>=2.9
PyJWT>=2.8
requests>=2.31
python-dotenv>=1.0
